// src/interfaces/consortiumAttributes.interface.ts
export interface ConsortiumAttributes {
  id: number;
  register_code: number;
  payment_cbu: number;
  payment_alias: string;
  payment_account_num: string;
  payment_bank: string;
  payment_holder: string;
  email: string;
  phone: string;
  address: string;

  //revisar con glootiex o robberinho
  units: number;
}